##################################################################### ## Modification 12/2015: Vegetation forcing and variable DZ in CLM ## #####################################################################
11 Dec. 2015

Basile HECTOR (BH)
- Laboratoire des Transferts en Hydrologie et Environnement (LTHE), Domaine universitaire BP 53, 38041 Grenoble Grenoble, France
- National Center for Space Studies (CNES), France
- basilehector@gmail.com

########################
## Vegetation forcing ##
########################

Added the option to force the vegetation in CLM (lai, sai, z0m, displa).

A new key is available in the tcl script: ForceVegetation. Default is False. If True, different input files are required depending on whether the meteorological CLM forcing is 1D or 3D:

- 1D forcing case: the user must provide 4 ascii text files (lai.dat, sai.dat, z0m.dat, displa.dat) which contain as many columns (time series) as IGBP vegetation classes (usually 18). The user also provides a vegetation map as a properly distributed pfb file (veg_map.pfb). Values are IGBP vegetation classes that are used to pass 2D maps of lai, sai, displa and z0m from ParFlow to CLM at each timestep.
The vegetation map could be later passed to CLM to remove the dependency on driver files.

- 3D forcing case: the user provides 4 properly distributed 3D (2D + time) files (LAI.pfb, SAI.pfb, Z0M.pfb and DISPLA.pfb) following the same rules as 3D meteorological forcing.

Added test case in the test/clm/ directory: clm_forc_veg.tcl
Added manual entries for this option.

Modified files:
- CLM.F90
- drv_getforce.F90
- solver_richards.c
- parflow_proto_f.h

Changes can be tracked by searching for "BH" flags

########################
## Variable DZ in CLM ##
########################
It has been observed that root fraction distribution in CLM was based on constant DZ, regardless of whether variable DZ was selected in ParFlow. Namely, the call to drv_clmini.F90 occurred before the variable DZ was defined in CLM. This is fixed in this version. The lines defining the root fraction have been copied in CLM.F90 after the variable DZ definition.

However the layer thicknesses and layer interfaces do not match between CLM and ParFlow, only center of cells location match. As a consequence, the volume of the column is not preserved. This has been corrected in this version, and the root fraction distribution is calculated based on layer interfaces which match ParFlow layer interfaces.

Added test case in the test/clm/ directory: clm_varDZ.tcl

Modified files:
- CLM.F90

Changes can be tracked by searching for "BH" flags


git-svn-id: https://parflow.svn.cloudforge.com/parflow/parflow/trunk@727 39af3fb0-ca4e-0410-adf9-b4ae76d04516