Implement netcdf-writer FlowVR module
Untested! Parflow does simulation work when it gets messages by e.g. the
newly implemented netcdf-writer FlowVR module. Thus the interface has
heavily changed. The commited state compiles if the right build-parameters
are choosen but is still untested as for a test run more programming work
has to be done.