Replace octree iteration with box iteration. (#148)
The octree loop iteration was replaced with iteration over a simple array of boxes that cover the geometry. For backward compatibility, the new clustering method can be turned off with the UseClustering input key.

The octree loop iteration was not very efficient. For example a flat plane of data (surface) was causing iterations into 1x1x1 leaf nodes in the octree due to power of 2 spatial division that is inherent to a standard octree representation of space. To compute the set of boxes, the Berger-Rigoutsos algorithm is run on a set of tags setup using the octree. This will compute a reasonably efficient set of boxes for iteration. For example, the top surface is now a single box if top is a uniform single geometry. Based on tests run on larger problem this is always more efficient. For very small problems the BR algorithm cost can be larger than loop savings.

This change should also be helpful for the work going with accelerators since the iteration will be over larger sized patches.