Add and simplify location methods
In addition to the "select()" method (which was renamed from
"get_selection_from_location()"), three new location methods
were added:

* value()
* doc()
* details()

All of the location methods also now function with either dot notation
or slash notation, due to a decorator that normalizes the location.
The decorator automatically assumes that the first string argument is
the location, and it normalizes it.

This also adds some tests to ensure things are working properly.

Note: we might be able to further simplify the logic in some of these
selection methods.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>