Update keys.rst
fixed bugs in irrigation documentation per user email (thanks Nick Jadallah)

The issue is a typo. ReadTheDocs has the irrigation type key with an 's' at the end but there should not be an 's' there. 

Currently, it shows this (wrong): 
<runname>.Solver.CLM.IrrigationTypes "Drip"  ## Python syntax

It should show this instead (correct):  
<runname>.Solver.CLM.IrrigationType "Drip"  ## Python syntax

I'm not sure who the best person is to make an edit to ReadTheDocs, but that might be good to change to save other people a headache. There's also the 
IrrigationStartTime/IrrigationStopTime typo that should also be fixed (float between 1-24 in GMT).