Use file hashes to trigger Github Actions cache builds (#625)
Previous method of using variable to force rebuilds was manual and
caused issues. Forks can not see the cache variable which also can cause
problems. Changed to use hash of the workflow files to trigger cache
builds. This is conservative and will cause some unnecessary builds but
workflow changes are relatively infrequent.