Refactor allocation of Overland variables (#619)
This PR addresses an issue brought up during the core ParFlow developer
meeting of June 9th, 2025.

In summary, there were identified a combined total of 16 variables in
the main `NlFunctionEval()` and `RichardsJacobianEval()` modules that
were allocated and deallocated at every invoke call of the modules,
which can happen multiple times in each nonlinear iteration.
Additionally, most of these variables are only used when a boundary
condition (BC) of the type OverlandFlow is active.

Therefore, this PR introduces a few updates:
1. Moves variable allocation outside of the iteration loop;
2. Only allocates variables when OverlandFlow-type BCs are active;
3. Achieves performance gains in solver time around 12% for problems
without OverlandFlow;
4. Achieves performance gains in solver time around 8% for problems with
OverlandFlow;
5. Achieves small memory savings for problems without OverlandFlow.

---------

Co-authored-by: Steven Smith <smith84@llnl.gov>