Alternative implementation with each patch having its own seepage key (#676)
This PR changes the Seepage Patch API so that each patch can define if
its a seepage patch separately, like this:

run.Patch.<patch_name>.BCPressure.Seepage = True.

The only changes are in the way keys are read from the API. The code
iterates over all patches and gathers which ones have both
OverlandKinematic and Seepage on. The underlying implementation for the
mixed BC remains exactly the same.

---------

Co-authored-by: George Artavanis <ga6@verde.princeton.edu>
Co-authored-by: Steven Smith <smith84@llnl.gov>