Feature/mixed surface bcs (#585)
This feature adds a mixed boundary condition capability to ParFlow. If the top boundary is set to OverlandKinematic and a pfsolid file is used, the solid file can have multiple patches with mixed boundary conditions. One patch is assigned to the overland kinematic seepage patches can be mixed in with the overland flow.   

Use the Seepage input key set to true on the seepage patches.  E.G.
      CONUS2.Patch.sink.BCPressure.Seepage = True

tagging @smithsg84 @lecondon to review.

---------

Co-authored-by: Reed <rmaxwell@mines.edu>
Co-authored-by: Steven Smith <smith84@llnl.gov>
Co-authored-by: gartavanis <32808713+gartavanis@users.noreply.github.com>
Co-authored-by: George Artavanis <ga6@verde.princeton.edu>